+++ Clockwork/Support/Vanilla/config.php
@@ -17,6 +17,26 @@ return [
 
 	/*
 	|------------------------------------------------------------------------------------------------------------------
+	| Features
+	|------------------------------------------------------------------------------------------------------------------
+	|
+	| You can enable or disable various Clockwork features here. Some features have additional settings (eg. slow query
+	| threshold for database queries).
+	|
+	*/
+
+	'features' => [
+
+		// Performance metrics
+		'performance' => [
+			// Allow collecting of client metrics. Requires separate clockwork-browser npm package.
+			'client_metrics' => isset($_ENV['CLOCKWORK_PERFORMANCE_CLIENT_METRICS']) ? $_ENV['CLOCKWORK_PERFORMANCE_CLIENT_METRICS'] : true
+		]
+
+	],
+
+	/*
+	|------------------------------------------------------------------------------------------------------------------
 	| Enable toolbar
 	|------------------------------------------------------------------------------------------------------------------
 	|